resolve Aaron's comments
