+++ src/pages/studyView/StudyViewUtils.tsx
@@ -17,15 +17,19 @@ import {
     GenePanelData,
     GenericAssayData,
     GenericAssayDataBin,
+    GenericAssayDataFilter,
     GenericAssayDataMultipleStudyFilter,
     GenomicDataBin,
     GenomicDataCount,
     MolecularDataMultipleStudyFilter,
     MolecularProfile,
     NumericGeneMolecularData,
+    Patient,
     PatientIdentifier,
+    PatientTreatmentRow,
     Sample,
     SampleIdentifier,
+    SampleTreatmentRow,
     StructuralVariantFilterQuery,
     StudyViewFilter,
 } from 'cbioportal-ts-api-client';
@@ -36,6 +40,7 @@ import {
     BinMethodOption,
     GenericAssayChart,
     GenomicChart,
+    SurvivalType,
     XvsYChartSettings,
     XvsYScatterChart,
     XvsYViolinChart,
@@ -88,11 +93,21 @@ import {
 import { getServerConfig } from 'config/config';
 import joinJsx from 'shared/lib/joinJsx';
 import { BoundType, NumberRange } from 'range-ts';
-import { ClinicalEventTypeCount } from 'cbioportal-ts-api-client/dist/generated/CBioPortalAPIInternal';
+import {
+    ClinicalEventTypeCount,
+    GenericAssayDataCountFilter,
+    GenericAssayDataCountItem,
+    GenomicDataCountFilter,
+    GenomicDataFilter,
+} from 'cbioportal-ts-api-client/dist/generated/CBioPortalAPIInternal';
 import { queryContainsStructVarAlteration } from 'shared/lib/oql/oqlfilter';
 import { toast } from 'react-toastify';
 import { useCallback } from 'react';
 import { MutationOptionConstants } from 'shared/constants';
+import { MolecularAlterationType_filenameSuffix } from 'shared/lib/StoreUtils';
+import { MultiSelectionTableRow } from './table/MultiSelectionTable';
+import Survival from 'pages/groupComparison/Survival';
+import { StructVarMultiSelectionTableRow } from './table/StructuralVariantMultiSelectionTable';
 
 // Cannot use ClinicalDataTypeEnum here for the strong type. The model in the type is not strongly typed
 export enum ClinicalDataTypeEnum {
@@ -4049,4 +4064,580 @@ function ToastSuppressor(props: any) {
             <a onClick={suppressToast}>Don't show this again</a>
         </>
     );
+}
+
+export function getDefaultClinicalDataBinFilter(
+    attribute: ClinicalAttribute
+): ClinicalDataBinFilter & {
+    showNA?: boolean | undefined;
+} {
+    return {
+        attributeId: attribute.clinicalAttributeId,
+        disableLogScale: false,
+        showNA: true,
+    } as ClinicalDataBinFilter & { showNA?: boolean };
+}
+
+export function addColorToCategories(
+    counts: ClinicalDataCount[],
+    attributeId: string,
+    getDisplayedValue?: (value: string) => string,
+    getDisplayedColor?: (value: string) => string
+): ClinicalDataCountSummary[] {
+    return getClinicalDataCountWithColorByClinicalDataCount(counts).map(
+        item => {
+            if (getDisplayedValue) {
+                item.displayedValue = getDisplayedValue(item.value);
+            }
+
+            if (getDisplayedColor) {
+                return {
+                    ...item,
+                    color: getDisplayedColor(item.value),
+                };
+            }
+
+            let colorMapKey = this.generateColorMapKey(attributeId, item.value);
+            // If the item doesn't has an assigned color
+            if (!this.chartItemToColor.has(colorMapKey)) {
+                // If the color has not been used
+                if (!this.chartToUsedColors.get(attributeId)?.has(item.color)) {
+                    this.chartItemToColor.set(colorMapKey, item.color);
+                    this.chartToUsedColors.get(attributeId)?.add(item.color);
+                } else {
+                    // Pick up a new color if the color has been used
+                    let d = {
+                        value: item.value,
+                        count: item.count,
+                    };
+                    let newColor = pickNewColorForClinicData(
+                        d,
+                        this.chartToUsedColors.get(attributeId) || new Set()
+                    );
+                    this.chartItemToColor.set(colorMapKey, newColor);
+                    this.chartToUsedColors.get(attributeId)?.add(newColor);
+                    item.color = newColor;
+                }
+                return item;
+            } else {
+                return {
+                    ...item,
+                    color: this.chartItemToColor.get(colorMapKey)!,
+                };
+            }
+        }
+    );
+}
+
+export async function invokeGenericAssayDataCount(
+    chartInfo: GenericAssayChart,
+    filters: StudyViewFilter
+) {
+    let result: GenericAssayDataCountItem[] = [];
+
+    result = await internalClient.fetchGenericAssayDataCountsUsingPOST({
+        genericAssayDataCountFilter: {
+            genericAssayDataFilters: [
+                {
+                    stableId: chartInfo.genericAssayEntityId,
+                    profileType: chartInfo.profileType,
+                } as GenericAssayDataFilter,
+            ],
+            studyViewFilter: filters,
+        } as GenericAssayDataCountFilter,
+    });
+
+    let data = result.find(d => d.stableId === chartInfo.genericAssayEntityId);
+    let counts: ClinicalDataCount[] = [];
+    let stableId: string = '';
+    if (data !== undefined) {
+        counts = data.counts.map(c => {
+            return {
+                count: c.count,
+                value: c.value,
+            } as ClinicalDataCount;
+        });
+        stableId = data.stableId;
+        if (!this.chartToUsedColors.has(stableId)) {
+            this.chartToUsedColors.set(stableId, new Set());
+        }
+    }
+
+    return addColorToCategories(counts, stableId);
+}
+
+export async function invokeGenomicDataCount(
+    chartInfo: GenomicChart,
+    filters: StudyViewFilter
+) {
+    let result = [];
+    let getDisplayedValue;
+    let params = {
+        genomicDataCountFilter: {
+            genomicDataFilters: [
+                {
+                    hugoGeneSymbol: chartInfo.hugoGeneSymbol,
+                    profileType: chartInfo.profileType,
+                } as GenomicDataFilter,
+            ],
+            studyViewFilter: filters,
+        },
+    } as any;
+
+    // mutation data counts (pie chart)
+    if (
+        chartInfo.profileType ===
+        MolecularAlterationType_filenameSuffix.MUTATION_EXTENDED
+    ) {
+        params = {
+            ...params,
+            $queryParameters: {
+                projection: 'SUMMARY',
+            },
+        };
+        result = await internalClient.fetchMutationDataCountsUsingPOST(params);
+        getDisplayedValue = transformMutatedType;
+    } else {
+        result = await internalClient.fetchGenomicDataCountsUsingPOST(params);
+        getDisplayedValue = getCNAByAlteration;
+    }
+
+    const data = result.find(
+        d =>
+            d.hugoGeneSymbol === chartInfo.hugoGeneSymbol &&
+            d.profileType === chartInfo.profileType
+    );
+
+    let counts: ClinicalDataCount[] = [];
+    let profileType: string = '';
+    if (data !== undefined) {
+        counts = data.counts.map(c => {
+            return {
+                count: c.count,
+                value: c.value,
+            } as ClinicalDataCount;
+        });
+        profileType = data.profileType;
+    }
+
+    return this.addColorToCategories(counts, profileType, getDisplayedValue);
+}
+
+export async function invokeMutationDataCount(
+    chartInfo: GenomicChart,
+    filters: StudyViewFilter,
+    profiledCases: number
+) {
+    const params = {
+        genomicDataCountFilter: {
+            genomicDataFilters: [
+                {
+                    hugoGeneSymbol: chartInfo.hugoGeneSymbol,
+                    profileType: chartInfo.profileType,
+                } as GenomicDataFilter,
+            ],
+            studyViewFilter: filters,
+        },
+        $queryParameters: {
+            projection: 'DETAILED',
+        },
+    } as any;
+
+    const result = await internalClient.fetchMutationDataCountsUsingPOST(
+        params
+    );
+
+    const data = result.find(
+        d =>
+            d.hugoGeneSymbol === chartInfo.hugoGeneSymbol &&
+            d.profileType === chartInfo.profileType
+    );
+
+    let counts: MultiSelectionTableRow[] = [];
+    if (data !== undefined) {
+        counts = data.counts.map(c => {
+            return {
+                uniqueKey: c.value,
+                label: c.label,
+                // "Altered" and "Profiled" really just mean
+                //  "numerator" and "denominator" in percent
+                //  calculation of table. Here, they mean
+                //  "# filtered samples in profile" and "# filtered samples overall"
+                numberOfAlteredCases: c.uniqueCount,
+                numberOfProfiledCases: profiledCases,
+                totalCount: c.count,
+            } as MultiSelectionTableRow;
+        });
+    }
+
+    return counts;
+}
+
+export async function getCustomChartDownloadData(
+    chartMeta: ChartMeta,
+    selectedSamples: Sample[],
+    selectedPatients: Patient[],
+    caseIdentifiers?: CustomChartIdentifierWithValue[]
+): Promise<string> {
+    return new Promise<string>(resolve => {
+        if (chartMeta && chartMeta.uniqueKey) {
+            let isPatientChart = chartMeta.patientAttribute;
+            let header = ['Study ID', 'Patient ID'];
+
+            if (!isPatientChart) {
+                header.push('Sample ID');
+            }
+            header.push(chartMeta.displayName);
+            let data = [header.join('\t')];
+            if (
+                chartMeta.uniqueKey ===
+                SpecialChartsUniqueKeyEnum.CANCER_STUDIES
+            ) {
+                data = data.concat(
+                    selectedSamples.map((sample: Sample) => {
+                        return [
+                            sample.studyId || Datalabel.NA,
+                            sample.patientId || Datalabel.NA,
+                            sample.sampleId || Datalabel.NA,
+                            sample.studyId,
+                        ].join('\t');
+                    })
+                );
+            } else if (!_.isEmpty(caseIdentifiers)) {
+                if (isPatientChart) {
+                    data = data.concat(
+                        selectedPatients.map((patient: Patient) => {
+                            let record = _.find(
+                                caseIdentifiers,
+                                (
+                                    caseIdentifier: CustomChartIdentifierWithValue
+                                ) => {
+                                    return (
+                                        caseIdentifier.studyId ===
+                                            patient.studyId &&
+                                        patient.patientId ===
+                                            caseIdentifier.patientId
+                                    );
+                                }
+                            );
+                            return [
+                                patient.studyId || Datalabel.NA,
+                                patient.patientId || Datalabel.NA,
+                                record === undefined ? 'NA' : record.value,
+                            ].join('\t');
+                        })
+                    );
+                } else {
+                    data = data.concat(
+                        selectedSamples.map((sample: Sample) => {
+                            let record = _.find(
+                                caseIdentifiers,
+                                (
+                                    caseIdentifier: CustomChartIdentifierWithValue
+                                ) => {
+                                    return (
+                                        caseIdentifier.studyId ===
+                                            sample.studyId &&
+                                        sample.sampleId ===
+                                            caseIdentifier.sampleId
+                                    );
+                                }
+                            );
+                            return [
+                                sample.studyId || Datalabel.NA,
+                                sample.patientId || Datalabel.NA,
+                                sample.sampleId || Datalabel.NA,
+                                record === undefined ? 'NA' : record.value,
+                            ].join('\t');
+                        })
+                    );
+                }
+            }
+
+            resolve(data.join('\n'));
+        } else {
+            resolve('');
+        }
+    });
+}
+
+export async function getScatterDownloadData(
+    chartInfo: XvsYScatterChart,
+    promise: MobxPromise<Sample[]>
+): Promise<string> {
+    const selectedSamples = await toPromise(promise);
+    const [xData, yData] = await Promise.all([
+        getSampleToClinicalData(selectedSamples, chartInfo.xAttr),
+        getSampleToClinicalData(selectedSamples, chartInfo.yAttr),
+    ]);
+    return generateXvsYScatterPlotDownloadData(
+        chartInfo.xAttr,
+        chartInfo.yAttr,
+        selectedSamples,
+        xData,
+        yData
+    );
+}
+
+export function getSurvivalDownloadData(
+    chartMeta: ChartMeta,
+    survivalPlots: SurvivalType[],
+    survivalDataMap: { [id: string]: ClinicalData[] },
+    selectedPatients: Patient[],
+    selectedPatientKeys: string[]
+): string {
+    const matchedPlot = _.find(
+        survivalPlots,
+        plot => plot.id === chartMeta.uniqueKey
+    );
+    if (matchedPlot && survivalDataMap) {
+        const data: string[] = [];
+
+        // find the unique clinical attribute ids
+        const uniqueClinicalAttributeIds = matchedPlot.associatedAttrs;
+
+        // add the header row
+        data.push(
+            ['Study ID', 'Patient ID', ...uniqueClinicalAttributeIds].join('\t')
+        );
+
+        // add the data rows
+        const selectedPatientMap = _.reduce(
+            selectedPatients,
+            (acc, next) => {
+                acc[next.uniquePatientKey] = next;
+                return acc;
+            },
+            {} as { [uniquePatientKey: string]: Patient }
+        );
+        selectedPatientKeys.forEach(uniquePatientKey => {
+            const clinicalDataList = survivalDataMap[uniquePatientKey];
+            const row: string[] = [];
+
+            if (clinicalDataList && clinicalDataList.length > 0) {
+                row.push(clinicalDataList[0].studyId || Datalabel.NA);
+                row.push(clinicalDataList[0].patientId || Datalabel.NA);
+                const keyed = _.keyBy(clinicalDataList, 'clinicalAttributeId');
+
+                _.each(uniqueClinicalAttributeIds, id => {
+                    row.push(
+                        keyed[id]
+                            ? keyed[id].value || Datalabel.NA
+                            : Datalabel.NA
+                    );
+                });
+            } else {
+                const selectedPatient = selectedPatientMap[uniquePatientKey];
+                if (selectedPatient) {
+                    row.push(selectedPatient.studyId || Datalabel.NA);
+                    row.push(selectedPatient.patientId || Datalabel.NA);
+
+                    _.each(uniqueClinicalAttributeIds, () => {
+                        row.push(Datalabel.NA);
+                    });
+                }
+            }
+
+            data.push(row.join('\t'));
+        });
+
+        return data.join('\n');
+    } else {
+        return '';
+    }
+}
+
+export async function getMutatedGenesDownloadData(
+    promise: MobxPromise<MultiSelectionTableRow[]>,
+    oncokbCancerGeneFilterEnabled: boolean
+): Promise<string> {
+    if (promise.result) {
+        let header = [
+            'Gene',
+            'MutSig(Q-value)',
+            '# Mut',
+            '#',
+            'Profiled Samples',
+            'Freq',
+        ];
+        if (oncokbCancerGeneFilterEnabled) {
+            header.push('Is Cancer Gene (source: OncoKB)');
+        }
+        let data = [header.join('\t')];
+        _.each(promise.result, (record: MultiSelectionTableRow) => {
+            let rowData = [
+                record.label,
+                record.qValue === undefined ? '' : getQValue(record.qValue),
+                record.totalCount,
+                record.numberOfAlteredCases,
+                record.numberOfProfiledCases,
+                getFrequencyStr(
+                    (record.numberOfAlteredCases /
+                        record.numberOfProfiledCases) *
+                        100
+                ),
+            ];
+            if (oncokbCancerGeneFilterEnabled) {
+                rowData.push(
+                    this.oncokbCancerGeneFilterEnabled
+                        ? record.isCancerGene
+                            ? 'Yes'
+                            : 'No'
+                        : 'NA'
+                );
+            }
+            data.push(rowData.join('\t'));
+        });
+        return data.join('\n');
+    } else return '';
+}
+
+export function getStructuralVariantGenesDownloadData(
+    promise:
+        | MobxPromise<MultiSelectionTableRow[]>
+        | MobxPromise<StructVarMultiSelectionTableRow[]>,
+    oncokbCancerGeneFilterEnabled: boolean
+): string {
+    if (promise.result) {
+        const header = [
+            'Gene',
+            '# Structural Variant',
+            '#',
+            'Profiled Samples',
+            'Freq',
+        ];
+        if (oncokbCancerGeneFilterEnabled) {
+            header.push('Is Cancer Gene (source: OncoKB)');
+        }
+        const data = [header.join('\t')];
+        _.each(promise.result, (record: MultiSelectionTableRow) => {
+            const rowData = [
+                record.label,
+                record.totalCount,
+                record.numberOfAlteredCases,
+                record.numberOfProfiledCases,
+                getFrequencyStr(
+                    (record.numberOfAlteredCases /
+                        record.numberOfProfiledCases) *
+                        100
+                ),
+            ];
+            if (oncokbCancerGeneFilterEnabled) {
+                rowData.push(
+                    oncokbCancerGeneFilterEnabled
+                        ? record.isCancerGene
+                            ? 'Yes'
+                            : 'No'
+                        : 'NA'
+                );
+            }
+            data.push(rowData.join('\t'));
+        });
+        return data.join('\n');
+    } else return '';
+}
+
+export async function getGenesCNADownloadData(
+    promise: MobxPromise<MultiSelectionTableRow[]>,
+    oncokbCancerGeneFilterEnabled: boolean
+): Promise<string> {
+    if (promise.result) {
+        let header = [
+            'Gene',
+            'Gistic(Q-value)',
+            'Cytoband',
+            'CNA',
+            'Profiled Samples',
+            '#',
+            'Freq',
+        ];
+        if (oncokbCancerGeneFilterEnabled) {
+            header.push('Is Cancer Gene (source: OncoKB)');
+        }
+        let data = [header.join('\t')];
+        _.each(promise.result, (record: MultiSelectionTableRow) => {
+            let rowData = [
+                record.label,
+                record.qValue === undefined ? '' : getQValue(record.qValue),
+                record.cytoband,
+                getCNAByAlteration(record.alteration!),
+                record.numberOfAlteredCases,
+                record.numberOfProfiledCases,
+                getFrequencyStr(
+                    (record.numberOfAlteredCases /
+                        record.numberOfProfiledCases) *
+                        100
+                ),
+            ];
+            if (oncokbCancerGeneFilterEnabled) {
+                rowData.push(
+                    oncokbCancerGeneFilterEnabled
+                        ? record.isCancerGene
+                            ? 'Yes'
+                            : 'No'
+                        : 'NA'
+                );
+            }
+            data.push(rowData.join('\t'));
+        });
+        return data.join('\n');
+    } else return '';
+}
+
+export async function getPatientTreatmentDownloadData(
+    promise: MobxPromise<PatientTreatmentRow[]>
+): Promise<string> {
+    if (promise.result) {
+        const header = ['Treatment', '#'];
+        let data = [header.join('\t')];
+        _.each(promise.result, (record: PatientTreatmentRow) => {
+            let rowData = [record.treatment, record.count];
+            data.push(rowData.join('\t'));
+        });
+        return data.join('\n');
+    } else return '';
+}
+
+export async function getSampleTreatmentDownloadData(
+    promise: MobxPromise<SampleTreatmentRow[]>
+): Promise<string> {
+    if (promise.result) {
+        const header = ['Treatment', 'Pre/Post', '#'];
+        let data = [header.join('\t')];
+        _.each(promise.result, (record: SampleTreatmentRow) => {
+            let rowData = [record.treatment, record.time, record.count];
+            data.push(rowData.join('\t'));
+        });
+        return data.join('\n');
+    } else return '';
+}
+
+export async function getMutationTypesDownloadData(
+    promise: MobxPromise<MultiSelectionTableRow[]>
+): Promise<string> {
+    if (promise.result) {
+        let header = [
+            'Mutation Event',
+            '# Mut',
+            '#',
+            'Profiled Samples',
+            'Freq',
+        ];
+        let data = [header.join('\t')];
+        _.each(promise.result, (record: MultiSelectionTableRow) => {
+            let rowData = [
+                record.label,
+                record.totalCount,
+                record.numberOfAlteredCases,
+                record.numberOfProfiledCases,
+                getFrequencyStr(
+                    (record.numberOfAlteredCases /
+                        record.numberOfProfiledCases) *
+                        100
+                ),
+            ];
+            data.push(rowData.join('\t'));
+        });
+        return data.join('\n');
+    } else return '';
 }
